requirements.txt
requests>=2.31
browser-cookie3>=0.19
pystray>=0.19.5
Pillow>=10.0
